[Refactor] relocate search login and update for new domain
