+++ src/main/java/com/capstone/favicon/dataset/dto/SearchDto.java
@@ -1,4 +1,4 @@
-package com.capstone.favicon.user.dto;
+package com.capstone.favicon.dataset.dto;
 
 import lombok.Getter;
