--- src/main/java/com/capstone/favicon/user/application/service/DataService.java
@@ -1,10 +1,0 @@
-package com.capstone.favicon.user.application.service;
-
-import com.capstone.favicon.user.domain.Data;
-
-import java.util.List;
-
-public interface DataService {
-    List<Data> search(String text);
-    List<Data> searchWithCategory(String text, String category);
-}
